Update README and add prefilled environments
To make installation of the package easier a number of prefilled .plist files have been created. A user may copy one of these files to their app target and the environment variables will be passed through to the test runner.

Instructions have been added to the README file for how to use these files. In addition to instructions for how to manually pass the variables through if they cannot use the plist files for some reason.